+++ src/include/utils/tuplestorenew.h
@@ -25,9 +25,9 @@ typedef struct NTupleStore NTupleStore;
 void ntuplestore_setinstrument(NTupleStore* ts, struct Instrumentation *ins);
 
 /* Tuple store method */
-extern NTupleStore *ntuplestore_create(int maxBytes);
-extern NTupleStore *ntuplestore_create_readerwriter(const char* filename, int maxBytes, bool isWriter);
-extern NTupleStore *ntuplestore_create_workset(workfile_set *workSet, int maxBytes);
+extern NTupleStore *ntuplestore_create(int64 maxBytes);
+extern NTupleStore *ntuplestore_create_readerwriter(const char* filename, int64 maxBytes, bool isWriter);
+extern NTupleStore *ntuplestore_create_workset(workfile_set *workSet, int64 maxBytes);
 extern bool ntuplestore_is_readerwriter_reader(NTupleStore* nts);
 extern void ntuplestore_reset(NTupleStore *ts);
 extern void ntuplestore_flush(NTupleStore *ts);
